Fix missing prototypes warning
Functions qoi_write_32 and qoi_read_32 should either have a prototype
or be declared static. Do the latter, as theses functions are defined
in a header and should only exists in the same compilation unit as the
file including qoi.h and defining QOI_IMPLEMENTATION, ie not exported.

By adding the static keyword in the function declaration the following
command doesn't raise a missing-prototypes warning, anymore:
    cc -Wmissing-prototypes -DQOI_IMPLEMENTATION -c -o qoi.o -xc qoi.h