Require smaller encodings when multiple encodings are possible
Before this commit, for some raw images, multiple qoi files can decode
to them. This commit now requires encoders to choose chunks in a
specific priority order -- an order that is already present in the
code. One of the effects of this is that larger encodings are not
valid when a smaller encoding is available. This commit also requires
encoders to use maximal munch for RUN chunks.

As an example, consider the raw RGB image with pixel values

0x000000 0x000000

Before this commit, an encoding (minus the header and footer) of

b11000000 b11000000

(i.e. two runs of length 1) would be valid, even though the following
encoding is shorter:

b11000001

(i.e. one run of length 2).

The priority order is RUN, INDEX, DIFF, LUMA, RGB, RGBA. This order
does slightly more than just decreasing encoding size, however. For
instance, for the raw RGB image

0x000000

this commit mandates the encoding use a RUN chunk rather than a DIFF
chunk, even though they both have the same size (b11000000
vs. b01101010). The priority order in this commit allows encoders to
choose a chunk type only once for each pixel and without needing any
lookahead.

One way to specify the priority order would be to add to the comments
on each chunk (other than RUN) a note about when the chunk must not be
emitted. This is the way the comment on INDEX chunks was written
before this commit. While doing so would create more eplicit rules, it
would be longer and more complex than simply specifying a chunk
priority order. As a result, this commit also removes that "must not
issue" directive on INDEX chunks -- that rule is implicit in the chunk
type priority order.